optimize the cs general project. Move the TxtLoader to a single file
