+++ unity/general/Src/WrapperGenerate.cs
@@ -5,62 +5,9 @@
 * This file is subject to the terms and conditions defined in file 'LICENSE', which is part of this source code package.
 */
 
-using System.IO;
-using System.Reflection;
-using System.Diagnostics;
 using Puerts;
 using System.Collections.Generic;
 using System;
-
-public class TxtLoader : ILoader
-{
-    public static string PathToBinDir(string appendix)
-    {
-        return Path.Combine(
-            System.Text.RegularExpressions.Regex.Replace(
-                Path.GetDirectoryName(Assembly.GetExecutingAssembly().GetName().CodeBase), "^file:(\\\\)?", ""
-            ),
-            appendix
-        );
-    }
-
-    private string root = PathToBinDir("../../Assets/Puerts/Src/Resources");
-    private string editorRoot = PathToBinDir("../../Assets/Puerts/Src/Editor/Resources");
-
-    public bool FileExists(string filepath)
-    {
-        return mockFileContent.ContainsKey(filepath) ||
-            File.Exists(Path.Combine(root, filepath)) ||
-            File.Exists(Path.Combine(editorRoot, filepath));
-    }
-
-    public string ReadFile(string filepath, out string debugpath)
-    {
-        debugpath = Path.Combine(root, filepath);
-        if (File.Exists(Path.Combine(editorRoot, filepath)))
-        {
-            debugpath = Path.Combine(editorRoot, filepath);
-        }
-
-        string mockContent;
-        if (mockFileContent.TryGetValue(filepath, out mockContent))
-        {
-            return mockContent;
-        }
-
-        using (StreamReader reader = new StreamReader(debugpath))
-        {
-            return reader.ReadToEnd();
-        }
-    }
-
-    private Dictionary<string, string> mockFileContent = new Dictionary<string, string>();
-    public void AddMockFileContent(string fileName, string content)
-    {
-        mockFileContent.Add(fileName, content);
-    }
-}
-
 
 [Configure]
 public class WrapperGenConfig
